refactor: checkpoints saving logic
